New naming convention for full-lex.sh tests
This script executes test_lex on all <example-name>.deca files in src/test/syntax.
The outputs are saved in files named <example-name>-lex.lis to differentiate them
from the output files of the syntax test which is executed on the same files.

A file named *_lexCorr.deca denotes an example that is syntaxically invalid but
lexically correct.